+++ docker/dev-env-clientside/stage_3_sdk/configs/version_of_dev_env.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+
+# This script is auto-generated during docker build
+# DO NOT EDIT MANUALLY
+
+PROJECT_VERSION="0.5.6"
+
+echo "Current Version of SDK Development Environment is: 0.5.6"